backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-groq
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
